assertions created

+++ src/main/java/org/example/Account.java
@@ -2,41 +2,47 @@ package org.example;
 
 public class Account {
     private Integer account;
-    private Float balance;
+    private Double balance;
     private String email;
     private String phone;
 
-    public Float getBalance() {
+    public Double getBalance() {
+
         return balance;
     }
 
-    // Setter
-    public void setBalance(Float newBalance) {
+    public void setBalance(Double newBalance) {
+
         this.balance = newBalance;
     }
 
     public Integer getAccount() {
+
         return account;
     }
 
-    // Setter
     public void setAccount(Integer newAccount) {
+
         this.account = newAccount;
     }
 
     public String getEmail() {
+
         return email;
     }
 
     public void setEmail(String newEmail) {
+
         this.email = newEmail;
     }
 
     public String getPhone() {
+
         return phone;
     }
 
     public void setPhone(String newPhone) {
+
         this.phone = newPhone;
     }
